feat: redo practice02

+++ practice02/run.sh
@@ -15,7 +15,7 @@ if [ -z "$WORK_DIR_HOST" ]; then
 fi
 
 STUDENT_ID_SUFFIX=$STUDENT_ID_SUFFIX
-INSTANCE_ID=orcl$STUDENT_ID_SUFFIX
+INSTANCE_ID=ORCL$STUDENT_ID_SUFFIX
 
 if [ -z "$ORACLE_PWD" ]; then
   ORACLE_PWD=$INSTANCE_ID
@@ -30,36 +30,18 @@ GROUP_ID=54321
 mkdir -p $WORK_DIR/data
 sudo chown $USER_ID:$GROUP_ID $WORK_DIR/data
 
-ORACLE_DATABASE_CONTAINER_NAME=25-database-homework-practice02
+echo "start building docker image for oracle database..."
+sudo docker build -t mhmzx/database-practice:oracle-database-$INSTANCE_ID \
+  --build-arg INSTANCE_ID=$INSTANCE_ID \
+  $WORK_DIR/docker
 
-echo "stopping existing instance of oracle database..."
-sudo docker stop $ORACLE_DATABASE_CONTAINER_NAME
-sleep 5
-
-echo "starting oracle database..."
-sudo docker run --rm -d \
-  --name $ORACLE_DATABASE_CONTAINER_NAME \
-  --network host \
-  -v $WORK_DIR_HOST/data:/opt/oracle/oradata \
-  -e TZ=$TZ \
-  -e ORACLE_PWD=$INSTANCE_ID \
-  -e ORACLE_SID=$INSTANCE_ID \
-  -e ORACLE_UNQNAME=$ORACLE_UNQNAME \
-  container-registry.oracle.com/database/enterprise:21.3.0.0
-
-until [ "$(docker inspect -f {{.State.Health.Status}} $ORACLE_DATABASE_CONTAINER_NAME)" = "healthy" ]; do
-  echo "waiting for oracle database..."
-  sleep 5
-done
-
-echo "oracle database started, connecting..."
-sleep 5
-sudo docker run --rm -it \
-  --network host \
-  --entrypoint sqlplus \
-  -e TZ=$TZ \
-  container-registry.oracle.com/database/enterprise:21.3.0.0 \
-  sys@//127.0.0.1:1521/$INSTANCE_ID as sysdba
-
-echo "stopping oracle database..."
-sudo docker stop $ORACLE_DATABASE_CONTAINER_NAME
+if [ $? == 0 ]; then
+  echo "starting oracle database..."
+  sudo docker run --rm -it \
+    --network host \
+    --hostname $INSTANCE_ID \
+    -v $WORK_DIR_HOST/data:/opt/oracledb \
+    -e TZ=$TZ \
+    -e ORACLE_PWD=$INSTANCE_ID \
+    mhmzx/database-practice:oracle-database-$INSTANCE_ID
+fi
